fixing the scripts for shifter's sake

+++ DQM/SiPixelCommon/test/submit.csh
@@ -4,8 +4,9 @@ setenv startdir `pwd`
 foreach file ( ${startdir}/Run_offline_DQM_*_cfg.py )
 
 mkdir ${startdir}/JOB_${i}
-
+if( -e submit_${i}.csh ) then
 rm submit_${i}.csh
+endif
 sed "s/NUM/${i}/" < submit_template.csh > submit_${i}.csh
 sed "s#CFGDIR#${startdir}#" < submit_${i}.csh > tmp.csh
 mv tmp.csh submit_${i}.csh
